Update users table migration to change telegram_id and chat_id types from INTEGER to BIGINT for improved compatibility with larger values.

+++ migrations/postgres/20251129181601_create_users_table.sql
@@ -2,8 +2,8 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS users (
     id UUID PRIMARY KEY NOT NULL,
-    telegram_id INTEGER UNIQUE NOT NULL,
-    chat_id INTEGER UNIQUE NOT NULL,
+    telegram_id BIGINT UNIQUE NOT NULL,
+    chat_id BIGINT UNIQUE NOT NULL,
     first_name TEXT,
     last_name TEXT,
     username TEXT NOT NULL,
